+++ src/args.rs
@@ -41,13 +41,17 @@ mod inner {
     #[derive(Clone, Debug)]
     pub struct Args {
         /// list of remaining arguments, for cheap cloning
-        items: Rc<[Arg]>,
+        pub(crate) items: Rc<[Arg]>,
         /// removed items, false - present, true - removed
         removed: Vec<bool>,
         remaining: usize,
 
+        #[doc(hidden)]
         /// Used to render an error message for [`parse`][crate::Parser::parse]
-        pub(crate) current: Option<usize>,
+        pub current: Option<usize>,
+        #[doc(hidden)]
+        /// "deeper" parser should win in or_else branches
+        pub depth: usize,
 
         /// used to pick the parser that consumes the left most item
         pub(crate) head: usize,
@@ -100,6 +104,7 @@ mod inner {
                 items: Rc::from(vec),
                 current: None,
                 head: usize::MAX,
+                depth: 0,
             }
         }
     }