+++ src/tests.rs
@@ -1303,7 +1303,7 @@ fn many_error_handling() {
 }
 
 #[test]
-fn failure_is_not_stupid() {
+fn failure_is_not_stupid_1() {
     let a = short('a').argument("A").from_str::<u32>();
     let b = pure(()).parse::<_, _, String>(|_| Err("nope".to_string()));
     let parser = construct!(a, b).to_options();
@@ -1313,4 +1313,113 @@ fn failure_is_not_stupid() {
         .unwrap_err()
         .unwrap_stderr();
     assert_eq!(res, "Couldn't parse: nope");
+}
+
+#[test]
+fn failure_is_not_stupid_2() {
+    let a = short('a').argument("A").from_str::<u32>();
+    let b = short('b').argument("B").from_str::<u32>();
+    let parser = construct!(a, b)
+        .parse::<_, (), String>(|_| Err("nope".to_string()))
+        .to_options();
+
+    let res = parser
+        .run_inner(Args::from(&["-a", "42", "-b", "42"]))
+        .unwrap_err()
+        .unwrap_stderr();
+    assert_eq!(res, "Couldn't parse: nope");
+}
+
+#[test]
+fn no_fallback_out_of_command_parser() {
+    let alt1 = positional("NAME").to_options().command("cmd");
+    let alt2 = pure(String::new());
+    let parser = construct!([alt1, alt2]).to_options();
+
+    let res = parser
+        .run_inner(Args::from(&["cmd"]))
+        .unwrap_err()
+        .unwrap_stderr();
+    assert_eq!(res, "Expected <NAME>, pass --help for usage information");
+
+    let res = parser.run_inner(Args::from(&["cmd", "a"])).unwrap();
+    assert_eq!(res, "a");
+
+    let res = parser.run_inner(Args::from(&[])).unwrap();
+    assert_eq!(res, "");
+}
+
+#[test]
+fn did_you_mean_switch() {
+    let a = short('f').long("flag").switch();
+    let b = short('p').long("plag").switch();
+    let parser = construct!([a, b]).to_options();
+
+    let res = parser
+        .run_inner(Args::from(&["--fla"]))
+        .unwrap_err()
+        .unwrap_stderr();
+    assert_eq!(res, "No such flag: `--fla`, did you mean `--flag`?");
+
+    let res = parser
+        .run_inner(Args::from(&["flag"]))
+        .unwrap_err()
+        .unwrap_stderr();
+    assert_eq!(res, "No such command: `flag`, did you mean `--flag`?");
+
+    let res = parser
+        .run_inner(Args::from(&["--pla"]))
+        .unwrap_err()
+        .unwrap_stderr();
+    assert_eq!(res, "No such flag: `--pla`, did you mean `--plag`?");
+
+    let res = parser
+        .run_inner(Args::from(&["--p"]))
+        .unwrap_err()
+        .unwrap_stderr();
+    assert_eq!(res, "No such flag: `--p`, did you mean `-p`?");
+}
+
+#[test]
+fn did_you_mean_req_flag() {
+    let parser = long("flag").req_flag(()).to_options();
+    let res = parser
+        .run_inner(Args::from(&["--fla"]))
+        .unwrap_err()
+        .unwrap_stderr();
+    assert_eq!(res, "Expected --flag, pass --help for usage information");
+}
+
+#[test]
+fn did_you_mean_argument() {
+    let parser = long("flag").argument("VAL").to_options();
+    let res = parser
+        .run_inner(Args::from(&["--fla"]))
+        .unwrap_err()
+        .unwrap_stderr();
+    assert_eq!(
+        res,
+        "Expected --flag VAL, pass --help for usage information"
+    );
+}
+
+#[test]
+fn did_you_mean_command() {
+    let parser = pure(())
+        .to_options()
+        .command("command")
+        .short('c')
+        .to_options();
+
+    let res = parser
+        .run_inner(Args::from(&["comman"]))
+        .unwrap_err()
+        .unwrap_stderr();
+    assert_eq!(res, "No such command: `comman`, did you mean `command`?");
+
+    let res = parser
+        .run_inner(Args::from(&["--comman"]))
+        .unwrap_err()
+        .unwrap_stderr();
+    assert_eq!(res, "No such flag: `--comman`, did you mean `command`?");
 }